remove socket-server and integrate it to main server backend
