+++ client/src/socket/socket.ts
@@ -1,7 +1,7 @@
 import { io } from "socket.io-client";
 import { ISocket } from "../types";
 
-const URL = import.meta.env.MODE === 'production' ? `${import.meta.env.VITE_SOCKET_URL}` : "http://localhost:7000";
+const URL = import.meta.env.MODE === 'production' ? `${import.meta.env.VITE_SOCKET_URL}` : "http://localhost:5000";
 
 const socket: ISocket = io(URL, {
   autoConnect: false,
