--- socket-server/docker-compose.yml
@@ -1,7 +1,0 @@
-version: "3"
-
-services:
-  redis:
-    image: redis:5
-    ports:
-      - "6379:6379"
