--- socket-server/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "socket-server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/cluster.ts",
-  "scripts": {
-    "start": "node dist/cluster.js",
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/cluster.js\"",
-    "build": "tsc"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@socket.io/sticky": "^1.0.3",
-    "dotenv": "^16.3.1",
-    "ioredis": "^5.3.2",
-    "socket.io": "^4.7.0",
-    "socket.io-redis": "^6.1.1"
-  },
-  "devDependencies": {
-    "@types/node": "^20.3.1",
-    "concurrently": "^8.2.0",
-    "nodemon": "^2.0.22",
-    "typescript": "^5.1.6"
-  }
-}
